requirements.txt
django
django-filter
djangorestframework
psycopg2-binary
gunicorn
dj_database_url
pytest
